Add publish, yank, unyank functions to index module.
These functions only deal with updating the index. For operations like
publish, there are certainly `.crate` files to be managed on disk, but that
isn't happening here. This code is only concerned with updating the
registry index itself.

Likely these operations should be done *as the last step* when
publishing since the index controls the visibility of the package to
clients.