Add http tests, formalize service config/settings.
The main aim here is to increase coverage since we took a hit with the
addition of all the untested web server related code.

This diff adds what the actix-web docs refer to as "integration" tests,
which are tests written from the point of view of a client sending
requests to the web server.

In order to implement these tests, some minor refactors had to be made
so our test service could be created consistently, each test running in
its own sandbox (as far as the file system access goes.

A new `Settings` struct was added to hold the paths for the crate
storage and index directories which can be constructed from env vars, or
manually. This struct is shared with handlers with the `Data` extractor.

This allows us to avoid repeatedly looking at env vars in code that will
be tested, while also giving us a single source of truth for these
paths.

Previously deeper env var reads were unwrapping/expecting since
it was assumed `main()` would have already ensured the vars were set.
This is not the case in tests, or at least it's antagonistic to use env
vars in this situation.

The tests for the publish endpoint use a payload captured via "spying"
on the requests sent to the service by cargo. [mitmproxy] was used to
capture the bytes for this request body.

With mitmproxy running, I ran `cargo publish` with the `CARGO_HTTP_PROXY`
env var set, then copied the "flow" file to the `test_data/` dir so we
could use `include_bytes!()` in our test code.

[mitmproxy]: https://mitmproxy.org/