Add git endpoints for index fetching.
The registry endpoints are all about modifying the index. These git
endpoints are all about fetching.

Cargo uses git to sync index data incrementally, so in order to give it
what it wants we need to implement the "fetch" portions of the "smart"
git server protocol as detailed here:

- <https://git-scm.com/docs/http-protocol#_discovering_references>
- <https://git-scm.com/docs/http-protocol#_smart_service_git_upload_pack>

With these endpoints in place, folks will configure cargo to use the
registry with:

```toml
[registries]
estuary = { index = "<scheme>://<host>:<port>/git/index" }
```

The `Procfile` was introduced to the project because we (prior to this
diff) needed to run `git daemon` to serve the git repo in addition to
the estuary server.

I've left the `Procfile` where it is, but removed the git server from it now
that it's redundant.

Who knows, the `Procfile` may still be of interest later on if/when other
services are required for estuary to operate.