Add structop, rename some config env vars.
Some of the config jargon was taken directly from cargo's docs on alt
registries, but could be confusing. The public facing flags and env vars
have been renamed to try and make their purpose more clear.

Ex: the "API URL" is hopefully more clear when named "base url" since it
acts as the prefix for other urls.

Also, while we're adding new deps, I ran a cargo update and specified
patch level versions on the pre-1.0 crates (and others that play fast
and lose with semver).