Crate detail, crate version list.
New handlers, templates and supporting code to render pages for crates.

The crate detail page will default to the highest version if a specific
version isn't named explicitly.

Templates for the site have been reorganized to extend a common base.

Currently, tailwindcss is being sourced from a cdn which is not ideal.
A js project will be added so we can tree-shake the styles properly
before the next release.

Crate detail pages are fashioned after those on crates.io, but sadly the
index data we're keeping does not include the readme for each release.
We'll need some extra storage of our own if we want to display more than
the dependencies. For now, the page stays roughly bare.