Remove `anyhow` adopt `thiserror`.
The motivation for this change is mainly to do with the fact `thiserror`
lets us chain errors together without boxing or otherwise losing the
typing.

When generating HTTP responses for errors, we need to know the types of
errors we're handling. Anyhow sort of sets us up to need to downcast
to get the type information back whereas thiserror retains the types.