Build tree-shaken styles using tailwindcss-cli.
Adds an npm package to track the package versions we need for building
styles. New tailwind css and config files are now used to produce a
"dist" stylesheet, which scans the rust/template sources for class names
to include in the output.

The stylesheet is exposed to the front-facting site via a dedicated
handler which uses `include_str!()` to bake the styles into the binary.

If we end up adding more static files, we may want to generalize this to
support mutliple files from a single handler.
Potentially we could use <https://crates.io/crates/include_dir> to
capture js/fonts/styles produced by our npm project.

A danger with this approach is accidentally releasing versions of
estuary with outdated dist styles. We'll have to consider how to avoid
this.