Display a list of crates on the landing page.
This replaces the old git reflog content with a list of crate names
linked to their crate detail pages (latest version).

We will absolutely want to get pagination back into this (currently
we're dumping out the full list) but I think we should wait until we've
got a database in place. A database will allow us to show N most
recently uploaded/modified crates in a much more practical way.

The crate list is built using a series of globs. This was out of
laziness on my part. I think we should be able to write a better
performing implementation without adding an extra dependency given a
little time and focus.